requirements.txt
GitPython==1.0.1
requests==2.9.1
